Various IntelliJ fixes

+++ java/se/sics/mspsim/chip/CC1101.java
@@ -296,7 +296,7 @@ public class CC1101 extends Radio802154 implements USARTListener {
                         break;
 
                 case CC1101_STX:
-            int len = (int) (0xff&txfifo.get(0));
+            int len = 0xff&txfifo.get(0);
             txFooterCountdown = 1 + len + 1/*len*/;
             if (DEBUG) {
                 System.out.println("TX started: len = " + len + ", txFooterCountdown = " + txFooterCountdown);
@@ -553,7 +553,7 @@ public class CC1101 extends Radio802154 implements USARTListener {
             if (txSendSynchByteCnt < NUM_PREAMBLE + NUM_SYNCH) {
                 txSendSynchByteCnt++;
                 if (rfListener != null) {
-                    rfListener.receivedByte((byte) (SYNCH_BYTE_LAST));
+                    rfListener.receivedByte(SYNCH_BYTE_LAST);
                 }
                 cpu.scheduleTimeEventMillis(sendEvent, getInterByteDelayMs());
 
@@ -776,7 +776,7 @@ public class CC1101 extends Radio802154 implements USARTListener {
 
         @Override
         public int getActiveFrequency() {
-                return (int) 0; /* Not implemented */
+                return 0; /* Not implemented */
         }
 
         @Override
